// src/Models/schemedetail.model.ts
// export class scheme {
//     detailId: number = 0;
//     //  schemeImage: string = '';
//     description: string = '';
//     minimumAmount: number = 0;
//     maximumAmount: number = 0;
//     minimumInvestTime: number = 0;
//     maximumInvestTime: number = 0;
//     //  issueDate: Date = new Date();
//     // maturityDate: Date = new Date();
//     minimumAge: number = 0;
//     maximumAge: number = 0;
//     profitRatio: number = 0;
//     registrationCommRatio: number = 0;
//     installmentCommRatio: number = 0;
//     insuranceSchemeId: number = 0;
//     status: boolean = true;
// }
export class scheme {
    detailId: number = 0;

    description: string = '';
    minimumAmount: number = 0;
    maximumAmount: number = 0;
    minimumInvestTime: number = 0;
    maximumInvestTime: number = 0;

    minimumAge: number = 0;
    maximumAge: number = 0;
    profitRatio: number = 0;
    registrationCommRatio: number=0;
    installmentCommRatio: number = 0;
    insuranceSchemeId: number = 0;
    status: boolean = true;
}